this time really using int16 in Int16Rates, thanx Felix

+++ sw/airborne/math/pprz_algebra_int.h
@@ -124,9 +124,9 @@ struct Int32Mat33 {
 
 /* Rotational speed                              */
 struct Int16Rates {
-  int32_t p;
-  int32_t q;
-  int32_t r;
+  int16_t p;
+  int16_t q;
+  int16_t r;
 };
 
 /* Rotational speed                              */
